+++ zwo/interpreter.py
@@ -1,5 +1,4 @@
 from collections import abc
-from dataclasses import dataclass
 
 from zwo.parser import BLOCK_T, PARAM_T, Range, Tag, VAL_T
 
@@ -17,21 +16,32 @@ def _check_keys(required: set[Tag], check_tags: abc.KeysView[Tag], block_tag: Ta
     return True
 
 
-@dataclass(slots=True)
 class ZWOMValidator:
     raw_blocks: list[BLOCK_T]
+    validated_blocks: list[BLOCK_T]
+    _ftp: int | None
 
-    _ftp: int | None = None
+    _in_repeat: bool
+    _n_repeats: int
 
-    def __post_init__(self) -> None:
-        self.validate_scanned()
+    __slots__ = ("raw_blocks", "validated_blocks", "_ftp", "_in_repeat", "_n_repeats")
 
-    def validate_scanned(self) -> None:
+    def __init__(self, raw_blocks: list[BLOCK_T]) -> None:
+        self.raw_blocks = raw_blocks
+        self._ftp = None
+        self.validated_blocks = self.validate_scanned()
+
+    def validate_scanned(self) -> list[BLOCK_T]:
         if Tag.META not in self.raw_blocks[0]:
             raise ZWOMValidationError("ZWOM file must begin with a META block")
 
         self.visit_meta_block(self.raw_blocks[0][Tag.META], Tag.META)
 
+        # To account for expansion of any chunk repeats we need to build a new list
+        # Initialize it with the META block since we skip it in the rest of the validation
+        validated_blocks: list[BLOCK_T] = [self.raw_blocks[0]]
+        repeat_blocks = []
+        self._in_repeat = False
         for block in self.raw_blocks[1:]:
             # Blocks only have one key, so we can dispatch validators using the first key
             block_tag = next(iter(block))
@@ -43,6 +53,10 @@ class ZWOMValidator:
                     self.visit_ramp_block(params, block_tag)
                 case Tag.INTERVALS:
                     self.visit_interval_block(params, block_tag)
+                case Tag.START_REPEAT:
+                    self.visit_start_repeat_block(params, block_tag)
+                case Tag.END_REPEAT:
+                    self.visit_end_repeat_block(params, block_tag)
                 case _:
                     raise ZWOMValidationError(f"Unknown workout tag: '{block_tag}'")
 
@@ -55,6 +69,27 @@ class ZWOMValidator:
                         self.visit_cadence(val, block_tag)
                     case _:
                         continue
+
+            if self._in_repeat:
+                # Don't include the repeat metablocks in the final output
+                if block_tag != Tag.START_REPEAT:
+                    repeat_blocks.append(block)
+            else:
+                # Check to see if we've just hit the END_REPEAT tag & dump the blocks accordingly
+                if repeat_blocks:
+                    validated_blocks.extend(repeat_blocks * self._n_repeats)
+                    repeat_blocks.clear()
+                    self._n_repeats = -1
+                else:
+                    # Don't include the repeat metablocks in the final output
+                    if block_tag != Tag.END_REPEAT:
+                        validated_blocks.append(block)
+
+        # Make sure the chunk repeat block was closed
+        if self._in_repeat:
+            raise ZWOMValidationError("START_REPEAT is missing a matching END_REPEAT.")
+
+        return validated_blocks
 
     def visit_meta_block(self, params: PARAM_T, block_tag: Tag) -> None:
         required_tags = {Tag.NAME, Tag.AUTHOR, Tag.DESCRIPTION}
@@ -86,6 +121,29 @@ class ZWOMValidator:
     def visit_interval_block(self, params: PARAM_T, block_tag: Tag) -> None:
         required_tags = {Tag.REPEAT, Tag.DURATION, Tag.POWER}
         _check_keys(required_tags, params.keys(), block_tag)
+
+    def visit_start_repeat_block(self, params: PARAM_T, block_tag: Tag) -> None:
+        if self._in_repeat:
+            raise ZWOMValidationError("Nested block chunk repetition is not supported.")
+
+        required_tags = {Tag.REPEAT}
+        _check_keys(required_tags, params.keys(), block_tag)
+
+        n_repeats = params[Tag.REPEAT]  # If we're here we know the key is there
+        if not isinstance(n_repeats, int):
+            raise ZWOMValidationError("START_REPEAT must have an integer REPEAT value.")
+
+        if n_repeats == 0:
+            raise ZWOMValidationError("REPEAT must be > 0.")
+
+        self._in_repeat = True
+        self._n_repeats = n_repeats
+
+    def visit_end_repeat_block(self, params: PARAM_T, block_tag: Tag) -> None:
+        if not self._in_repeat:
+            raise ZWOMValidationError("Missing opening START_REPEAT block.")
+
+        self._in_repeat = False
 
     def visit_power(self, power_spec: VAL_T) -> None:
         # Validate that an FTP is set in order to use absolute watts